requirements.txt
numpy==2.2.6
numpy-quaternion==2024.0.8
matplotlib==3.10.3

# matplotlib should install these but in case it doesn't uncomment below:
#contourpy==1.3.2
#cycler==0.12.1
#fonttools==4.58.0
#kiwisolver==1.4.8
#packaging==25.0
#pillow==11.2.1
#pyparsing==3.2.3
#python-dateutil==2.9.0.post0
#six==1.17.0

# numpy-quaternion should install this but in case it doesn't uncomment below:
#scipy==1.15.3